Updated user related data models and tests to match latest GitLab API specs (#349).
