+++ src/main/java/org/gitlab4j/api/models/Member.java
@@ -5,14 +5,10 @@ import java.util.Date;
 
 import org.gitlab4j.api.utils.JacksonJson;
 
-public class Member {
+public class Member extends AbstractUser<Member> {
 
     private AccessLevel accessLevel;
-    private Date createdAt;
-    private Integer id;
-    private String name;
-    private String state;
-    private String username;
+    private Date expiresAt;
 
     public AccessLevel getAccessLevel() {
         return this.accessLevel;
@@ -22,44 +18,12 @@ public class Member {
         this.accessLevel = accessLevel;
     }
 
-    public Date getCreatedAt() {
-        return this.createdAt;
+    public Date getExpiresAt() {
+        return this.expiresAt;
     }
 
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public Integer getId() {
-        return this.id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getState() {
-        return this.state;
-    }
-
-    public void setState(String state) {
-        this.state = state;
-    }
-
-    public String getUsername() {
-        return this.username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
+    public void setExpiresAt(Date expiresAt) {
+        this.expiresAt = expiresAt;
     }
 
     public Member withAccessLevel(AccessLevel accessLevel) {
@@ -67,23 +31,8 @@ public class Member {
         return this;
     }
 
-    public Member withCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-        return this;
-    }
-
-    public Member withId(Integer id) {
-        this.id = id;
-        return this;
-    }
-
-    public Member withState(String state) {
-        this.state = state;
-        return this;
-    }
-
-    public Member withUsername(String username) {
-        this.username = username;
+    public Member withExpiresAt(Date expiresAt) {
+        this.expiresAt = expiresAt;
         return this;
     }
