+++ src/main/java/org/gitlab4j/api/models/Owner.java
@@ -1,41 +1,4 @@
 package org.gitlab4j.api.models;
 
-import java.util.Date;
-
-import org.gitlab4j.api.utils.JacksonJson;
-
-public class Owner {
-
-    private Date createdAt;
-    private Integer id;
-    private String name;
-
-    public Date getCreatedAt() {
-        return this.createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public Integer getId() {
-        return this.id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return (JacksonJson.toJsonString(this));
-    }
+public class Owner extends AbstractUser<Owner> {
 }
