+++ src/test/resources/org/gitlab4j/api/user.json
@@ -8,6 +8,9 @@
   "web_url": "http://localhost:3000/john_smith",
   "created_at": "2012-05-23T08:00:58Z",
   "is_admin": false,
+  "bio": "",
+  "location": "",
+  "public_email": "john@example.com",
   "skype": "",
   "linkedin": "",
   "twitter": "",
@@ -29,5 +32,7 @@
   "can_create_project": true,
   "two_factor_enabled": true,
   "external": false,
-  "shared_runners_minutes_limit": 133
+  "private_profile": false,
+  "shared_runners_minutes_limit": 133,
+  "extra_shared_runners_minutes_limit": 133
 }
